int64 and related link

+++ kafkarestv3/model_consumer_group_lag_summary_data.go
@@ -20,8 +20,8 @@ type ConsumerGroupLagSummaryData struct {
 	MaxLagClientId    string           `json:"max_lag_client_id"`
 	MaxLagTopicName   string           `json:"max_lag_topic_name"`
 	MaxLagPartitionId int32            `json:"max_lag_partition_id"`
-	MaxLag            int32            `json:"max_lag"`
-	TotalLag          int32            `json:"total_lag"`
+	MaxLag            int64            `json:"max_lag"`
+	TotalLag          int64            `json:"total_lag"`
 	MaxLagConsumer    Relationship     `json:"max_lag_consumer"`
 	MaxLagPartition   Relationship     `json:"max_lag_partition"`
 }
